add SnackBar to common
