+++ components/AdminPage/MenuManagement/RemoveItemFromCategory.tsx
@@ -5,20 +5,14 @@ import {
   Backdrop,
   Typography,
   Button,
-  Snackbar,
 } from '@material-ui/core'
 import { Category, MenuItem } from '@/models/place'
 import { useStyles } from '@/styles/modal'
 import 'firebase/storage'
 import * as updateService from '@/firebase/updateDocument'
-import { SnackbarOrigin } from '@material-ui/core/Snackbar'
-import Alert from '@material-ui/lab/Alert'
 import type { Color } from '@material-ui/lab/Alert'
 import { useState } from 'react'
-
-interface State extends SnackbarOrigin {
-  open: boolean
-}
+import SnackBar from '@/components/common/SnackBar'
 
 interface Props {
   categories: Category
@@ -41,26 +35,17 @@ const RemoveItemFromCategory = ({
 }: Props) => {
   const classes = useStyles()
 
-  const [state, setState] = useState<State>({
-    open: false,
-    vertical: 'top',
-    horizontal: 'center',
-  })
-  const { vertical, horizontal, open } = state
   const [message, setMessage] = useState({
     text: '',
     severity: 'error' as Color,
+    open: false,
   })
   const handleOpenAlert = (text: string, severity: Color) => {
-    setState({ ...state, open: true })
     setMessage({
       text: text,
       severity: severity,
+      open: true,
     })
-  }
-
-  const handleClose = () => {
-    setState({ ...state, open: false })
   }
 
   const handleRemove = () => {
@@ -78,17 +63,7 @@ const RemoveItemFromCategory = ({
 
   return (
     <>
-      <Snackbar
-        anchorOrigin={{ vertical, horizontal }}
-        autoHideDuration={1000}
-        open={open}
-        key={vertical + horizontal}
-        onClose={handleClose}
-      >
-        <Alert variant="filled" severity={message.severity}>
-          {message.text}
-        </Alert>
-      </Snackbar>
+      <SnackBar message={message} setMessage={setMessage} />
       <Modal
         className={classes.modal}
         open={openModalRemove}
